+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * Add `ManifestPushOptsBuilder::tls_verify`
 * Add `ContainerRestoreOptsBuilder::pod`
 * Add `ImagesRemoveOptsBuilder::ignore`
+* Add `ImagesRemoveOptsBuilder::lookup_manifest`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync